seedDemoData: extract DemoContentBuilder helper
Move pure content/data methods (breed/category/topic data, breed
article rendering, template wikitext) from the maintenance script
into a new final class SMW\Maintenance\DemoContentBuilder. The script
now holds a private DemoContentBuilder field and delegates to it for
all wikitext generation; orchestration (page edits, deletions, job
queue) stays on the script.

Drop the dead Main Page step. The script never updates Main Page on
any non-fresh wiki because isSeedPage() rejects pre-existing pages,
and Main Page is auto-created by every fresh MediaWiki install. Its
two #ask queries will be re-introduced on a dedicated SMW Query
Examples page in a follow-up commit.

No behavior change beyond the Main Page removal: rendered output for
breed pages, categories, topic pages, and the template is byte-
identical against a sampled subset (Border Collie, Labrador
Retriever, Chihuahua, Template:Dog breed, Category:Herding Group,
Dog Training).
